DD-56-refactor: paginationUtils 적용하여 코드 간소화
